requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-unfold==0.49.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
git-filter-repo==2.47.0
mysqlclient==2.2.7
PyJWT==2.10.1
python-dotenv==1.1.0
python-http-client==3.3.7
sendgrid==6.11.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tzdata==2025.1
vulture==2.14
